pg-plus/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2,<3.0
# Alternatively, install PyMySQL and update ENGINE accordingly.
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0
django-celery-beat>=2.6,<3.0
gunicorn>=21.2,<22.0
uvicorn[standard]>=0.27,<0.28
